:hammer: Break down system tests into smaller units
